Create char_detected
automatic counting of macroscopic particles of coal in peat. hardcode, need to finalise to user use of one function and get not the radius but the length and width of the particle. 

Then make functions for particle counting, histogram creation for boundary definition, area counting, particle classification (it's a couple of hours).